requirements.txt
# For tartans
requests==2.18.4
Pillow==5.1.0

# my scripts
Fabric3==1.14.post1

